+++ jdcrafte/CHANGELOG.md
@@ -1,1 +1,5 @@
 ### Unreleased changes
+
+#### New Features
+
+* Added the feeding trough, a block that fills up as feed is added.